feat(entity): add permission checks + use checks on delete

+++ packages/backend/src/application/services/age/age-service.ts
@@ -150,12 +150,23 @@ export const buildAgeService = ({ ageRepository }: AgeServiceDependencies) => {
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser?.role !== "admin") {
+    if (!loggedUser.permissions.age.canDelete) {
       const existingData = await ageRepository.findAgeById(id);
 
       if (existingData?.ownerId !== loggedUser?.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isAgeUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedAge = await ageRepository.deleteAgeById(id);
