+++ packages/backend/src/application/services/department/department-service.ts
@@ -194,12 +194,23 @@ export const buildDepartmentService = ({
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser?.role !== "admin") {
+    if (!loggedUser.permissions.department.canDelete) {
       const existingData = await departmentRepository.findDepartmentById(id);
 
       if (existingData?.ownerId !== loggedUser?.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isDepartmentUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedDepartment = await departmentRepository.deleteDepartmentById(id);
