+++ packages/backend/src/application/services/distance-estimate/distance-estimate-service.ts
@@ -150,12 +150,23 @@ export const buildDistanceEstimateService = ({ distanceEstimateRepository }: Dis
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser?.role !== "admin") {
+    if (!loggedUser.permissions.distanceEstimate.canDelete) {
       const existingData = await distanceEstimateRepository.findDistanceEstimateById(id);
 
       if (existingData?.ownerId !== loggedUser?.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isDistanceEstimateUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedDistanceEstimate = await distanceEstimateRepository.deleteDistanceEstimateById(id);
