+++ packages/backend/src/application/services/locality/locality-service.ts
@@ -187,12 +187,23 @@ export const buildLocalityService = ({ localityRepository, inventoryRepository }
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser?.role !== "admin") {
+    if (!loggedUser.permissions.locality.canDelete) {
       const existingData = await localityRepository.findLocalityById(id);
 
       if (existingData?.ownerId !== loggedUser?.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isLocalityUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedLocality = await localityRepository.deleteLocalityById(id);
