+++ packages/backend/src/application/services/sex/sex-service.ts
@@ -150,12 +150,23 @@ export const buildSexService = ({ sexRepository }: SexServiceDependencies) => {
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser?.role !== "admin") {
+    if (!loggedUser.permissions.sex.canDelete) {
       const existingData = await sexRepository.findSexById(id);
 
       if (existingData?.ownerId !== loggedUser?.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isSexUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedSex = await sexRepository.deleteSexById(id);
