+++ packages/backend/src/application/services/species-class/species-class-service.ts
@@ -187,12 +187,23 @@ export const buildSpeciesClassService = ({ classRepository, speciesRepository }:
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser.role !== "admin") {
+    if (!loggedUser.permissions.speciesClass.canDelete) {
       const existingData = await classRepository.findSpeciesClassById(id);
 
       if (existingData?.ownerId !== loggedUser.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isSpeciesClassUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedClass = await classRepository.deleteSpeciesClassById(id);
