+++ packages/backend/src/application/services/weather/weather-service.ts
@@ -168,12 +168,23 @@ export const buildWeatherService = ({ weatherRepository }: WeatherServiceDepende
     }
 
     // Check that the user is allowed to modify the existing data
-    if (loggedUser.role !== "admin") {
+    if (!loggedUser.permissions.weather.canDelete) {
       const existingData = await weatherRepository.findWeatherById(id);
 
       if (existingData?.ownerId !== loggedUser.id) {
         return err("notAllowed");
       }
+    }
+
+    const isEntityUsedResult = await isWeatherUsed(`${id}`, loggedUser);
+
+    if (isEntityUsedResult.isErr()) {
+      return err(isEntityUsedResult.error);
+    }
+
+    const isEntityUsed = isEntityUsedResult.value;
+    if (isEntityUsed) {
+      return err("isUsed");
     }
 
     const deletedWeather = await weatherRepository.deleteWeatherById(id);
